fix(tanstack-start): persist parsed locale cookie
