+++ .changeset/tanstack-locale-cookie.md
@@ -1,0 +1,5 @@
+---
+"gt-tanstack-start": patch
+---
+
+Persist TanStack Start locale resolution through the browser locale cookie.
